Moves types to separate file
